CU-2e77a31: Revert CAT and CDB abstraction
