+++ medcat/cat.py
@@ -40,7 +40,6 @@ from medcat.ner.transformers_ner import TransformersNER
 from medcat.utils.saving.serializer import SPECIALITY_NAMES, ONE2MANY
 from medcat.stats.stats import get_stats
 from medcat.utils.filters import set_project_filters
-from medcat.base import CATBase
 
 
 logger = logging.getLogger(__name__) # separate logger from the package-level one
@@ -49,7 +48,7 @@ logger = logging.getLogger(__name__) # separate logger from the package-level on
 HAS_NEW_SPACY = has_new_spacy()
 
 
-class CAT(CATBase):
+class CAT(object):
     """The main MedCAT class used to annotate documents, it is built on top of spaCy
     and works as a spaCy pipline. Creates an instance of a spaCy pipline that can
     be used as a spacy nlp model.
@@ -82,6 +81,7 @@ class CAT(CATBase):
         >>> spacy_doc = cat("Put some text here")
         >>> print(spacy_doc.ents) # Detected entities
     """
+    DEFAULT_MODEL_PACK_NAME = "medcat_model_pack"
 
     def __init__(self,
                  cdb: CDB,
@@ -216,7 +216,7 @@ class CAT(CATBase):
             version.medcat_version = __version__
             logger.warning("Please consider updating [description, performance, location, ontology] in cat.config.version")
 
-    def create_model_pack(self, save_dir_path: str, model_pack_name: str = CATBase.DEFAULT_MODEL_PACK_NAME, force_rehash: bool = False,
+    def create_model_pack(self, save_dir_path: str, model_pack_name: str = DEFAULT_MODEL_PACK_NAME, force_rehash: bool = False,
             cdb_format: str = 'dill') -> str:
         """Will crete a .zip file containing all the models in the current running instance
         of MedCAT. This is not the most efficient way, for sure, but good enough for now.
@@ -1552,6 +1552,14 @@ class CAT(CATBase):
             for doc in docs:
                 if hasattr(doc, "text"):
                     logger.warning("%s...", doc.text[:50])
+
+    @staticmethod
+    def _get_doc_annotations(doc: Doc):
+        if type(doc['annotations']) == list:  # type: ignore
+            return doc['annotations']  # type: ignore
+        if type(doc['annotations']) == dict:  # type: ignore
+            return doc['annotations'].values()  # type: ignore
+        return None
 
     def destroy_pipe(self):
         self.pipe.destroy()
